购物车，收藏，评论，图片

+++ src/main/java/com/lxlt/bean/Cart.java
@@ -20,7 +20,7 @@ public class Cart {
 
     private BigDecimal price;
 
-    private Short number;
+    private Integer number;
 
     private String[] specifications;
 
@@ -92,11 +92,11 @@ public class Cart {
         this.price = price;
     }
 
-    public Short getNumber() {
+    public Integer getNumber() {
         return number;
     }
 
-    public void setNumber(Short number) {
+    public void setNumber(Integer number) {
         this.number = number;
     }
